Move subclass_attention_backend and subclass_attention_backend_with_overrides
