+++ vllm/model_executor/layers/attention/encoder_only_attention.py
@@ -13,8 +13,6 @@ from vllm.v1.attention.backend import (
     AttentionMetadata,
     AttentionType,
     CommonAttentionMetadata,
-)
-from vllm.v1.attention.backends.utils import (
     subclass_attention_backend,
 )
 from vllm.v1.attention.selector import get_attn_backend
